Bugfix: HTML pipeline empty content handling
Content can be directly provided to the html pipeline;
in this mode the pipline does not make an http request
to fetch the data, instead the content provided is used.

Until this commit, the pipeline would disregard the content
passed to it, if the content is an empty string.
In this case an http request would be performed to fetch
the data.

This behaviour is fixed by this commit