Bugfix: Markdown stages in html pipe should not an extra wrapping div
This wrapping div would be added when there was more then one
element in the markdown (which is the usual case) by hyperscript
because hyperscript has no concept of multiple elements in the
root node.

Our code came to expect this wrapping which led to markdowns with
a single element being rendered as an empty document in some
cases (the single element would be deleted during the emit-html
when generating .content.children).

This commit fixes this issue by using hast-util-to-html instead of
hyperscript for the hast -> html conversion which does indeed have
a concept of multiple elements at the root node.

Fixes #157