feat(pipeline): Introduced explicit, named extension points
The previous iteration of pipeline extensibility exposed an implementation detail (the function
name) to enable extensibility, implicitly making each step in the pipeline a potential extension
point, and worse, part of the API contract. This change introduces an explicit extension point
marker, called `ext`. It can be used in two ways: (a) by setting the `ext` property of the function
that you want to make an extension point to the name of the extension point (b) by calling the
`ext(name)` chain method on `Pipeline`, which will set the extension point name for the last
function added to the pipeline.