build(dependencies): npm audit fix
Tidelift reported a dependency warning on the branch, which happens typically when you introduce a bad (insecure, deprecated or incompatibly licensed) dependency in `package.json`, but in this case it is due to the `eslint-utils` dependency, which has been marked as insecure and pinned to a higher (fixed) version in `package-lock.json` in the `master` branch. By simply running `npm install` in this branch or basing this branch on an older version of the `master` branch, which did not have the fix applied yet, you are reintroducing it, which is why Tidelift is complaining. Fortunately, the fix is easy: just run `npm audit fix` and commit the changed `package-lock.json`. /cc @MarquiseRosier @rofe