fix(embed): fix external embed service URL construction
the embed service expects the to-be-embedded URL to be passed after a trailing slash. The code did
not add the trailing slash, making it impossible for Adobe I/O Runtime to find the correct action to
execute. This change adds the correct slash again.

fixes https://github.com/adobe/helix-embed/issues/145 fixes #559