fix(all pipes): get-metadata is not map-enumeration stable
Force alphabetical sorting of types in case of collisions.

BREAKING CHANGE: Rather than relying on node's property enumeration to sort the
types for `is-…` types in the metadata, we are now enforcing a stable sort
order based on alphabetical ordering in case of count collisions.
The result is that there may be inversions of type order:
- text=1, heading=1: `is-text-heading` => `is-heading-text`
- text=2, heading=1: `is-text-heading`
- text=2, heading=1, image=1, list=1: `is-text-list-image` => `is-text-heading-image`

fix #669